Add noindex to hidden track pages
